+++ libReadFluka/src/UsxSuw.cxx
@@ -176,7 +176,7 @@ const char* UsxSuw::GetXtitle() const
 
 const char* UsxSuw::GetYtitle() const
 {
-  if (abs((float)GetType())<=1) return "#Omega [sr]";
+  if (abs(GetType())<=1) return "#Omega [sr]";
   else return "lg(#Omega/1 sr)";
 }
